+++ frontend/src/components/atoms/Icons/ClientsIcon.tsx
