+++ frontend/src/components/atoms/Icons/Logo.tsx
