+++ frontend/src/components/atoms/Icons/LogoClient.tsx
