+++ frontend/src/components/atoms/Icons/LogoEntreprise.tsx
