+++ frontend/src/components/atoms/Icons/LogoMark.tsx
