+++ frontend/src/components/atoms/Icons/PlusIcon.tsx
