+++ frontend/src/components/atoms/Icons/ProjectsIcon.tsx
