+++ frontend/src/components/atoms/Icons/searchIcon.tsx
