+++ frontend/src/components/atoms/illustrations/AuthIllus.tsx
