+++ frontend/src/components/atoms/illustrations/Error404.tsx
