+++ frontend/src/components/atoms/illustrations/cardBackground.tsx
