+++ frontend/src/components/organisms/DisplayCards.tsx
@@ -1,5 +1,5 @@
 import { CardVariant } from "./Cards";
-import { Project, Deliverable, Task } from "../__generated__/graphql-types";
+import { Project, Deliverable, Task } from "../../__generated__/graphql-types";
 import { ReactNode } from "react";
 
 interface DisplayCardsProps {
