+++ frontend/src/components/organisms/Navigation.tsx
@@ -5,7 +5,7 @@ import PlusIcon from "../atoms/Icons/PlusIcon";
 import ClientsIcon from "../atoms/Icons/ClientsIcon";
 import SettingsIcon from "../atoms/Icons/SettingsIcon";
 import LogoEntrepriseIcon from "../atoms/Icons/LogoEntreprise";
-import { useRoleTheme } from "../context/roleThemeContext";
+import { useRoleTheme } from "../../context/roleThemeContext";
 import LogoClientIcon from "../atoms/Icons/LogoClient";
 
 export default function Navigation() {
