+++ frontend/src/pages/Settings.tsx
@@ -1,4 +1,4 @@
-import Navigation from "../components/Navigation";
+import Navigation from "../components/organisms/Navigation";
 
 export default function Settings() {
   return (
